// src/generated/user-api/model/platformInvitationDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * User Service API
 * User service endpoints (dev docs via springdoc).
 * OpenAPI spec version: v1
 */
import type { PlatformInvitationDtoInvitationStatus } from './platformInvitationDtoInvitationStatus';
import type { PlatformInvitationDtoRoleType } from './platformInvitationDtoRoleType';

export interface PlatformInvitationDto {
  id?: number;
  createdByUserId?: number;
  inviteeEmail?: string;
  createdAt?: string;
  expiresAt?: string;
  confirmedAt?: string;
  invitationStatus?: PlatformInvitationDtoInvitationStatus;
  roleType?: PlatformInvitationDtoRoleType;
}